refactor(Supplement): supplement controller, entity 디렉토리 위치 변경 (DDD)
