+++ src/main/java/FitMate/FitMateBackend/chanhaleWorking/dto/SupplementListDto.java
@@ -3,6 +3,7 @@ package FitMate.FitMateBackend.chanhaleWorking.dto;
 import FitMate.FitMateBackend.cjjsWorking.service.storageService.S3FileService;
 import FitMate.FitMateBackend.consts.ServiceConst;
 import FitMate.FitMateBackend.domain.supplement.*;
+import FitMate.FitMateBackend.supplement.entity.Supplement;
 import lombok.Data;
 import lombok.NoArgsConstructor;
