+++ src/main/java/FitMate/FitMateBackend/cjjsWorking/dto/myfit/mySupplement/MySupplementReadAllResponse.java
@@ -2,7 +2,6 @@ package FitMate.FitMateBackend.cjjsWorking.dto.myfit.mySupplement;
 
 import FitMate.FitMateBackend.domain.myfit.MySupplement;
 import FitMate.FitMateBackend.domain.routine.Routine;
-import FitMate.FitMateBackend.domain.supplement.Supplement;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
