+++ src/main/java/FitMate/FitMateBackend/domain/recommendation/SupplementRecommendation.java
@@ -1,19 +1,15 @@
 package FitMate.FitMateBackend.domain.recommendation;
 
-import FitMate.FitMateBackend.consts.ServiceConst;
 import FitMate.FitMateBackend.domain.BodyData;
 import FitMate.FitMateBackend.domain.EnglishPurpose;
 import FitMate.FitMateBackend.domain.Purpose;
 import FitMate.FitMateBackend.domain.User;
-import FitMate.FitMateBackend.domain.supplement.Supplement;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @Entity
 @Getter
