+++ src/main/java/FitMate/FitMateBackend/domain/supplement/Protein.java
@@ -2,6 +2,7 @@ package FitMate.FitMateBackend.domain.supplement;
 
 import FitMate.FitMateBackend.chanhaleWorking.form.supplement.SupplementForm;
 import FitMate.FitMateBackend.consts.ServiceConst;
+import FitMate.FitMateBackend.supplement.entity.Supplement;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import lombok.Getter;
